knowledge-base-flask/knowledge-base-flask/requirements.txt
aiohappyeyeballs==2.4.3
aiohttp==3.10.10
aiosignal==1.3.1
annotated-types==0.7.0
anyio==4.9.0
asgiref==3.8.1
attrs==24.2.0
azure-core==1.31.0
azure-identity==1.19.0
black==23.12.1
blinker==1.7.0
cachetools==5.5.0
certifi==2023.11.17
cffi==1.17.1
charset-normalizer==3.4.0
click==8.1.7
croniter==3.0.3
cryptography==43.0.1
distlib==0.3.8
distro==1.9.0
docstring_parser==0.16
ecdsa==0.19.0
elastic-transport==8.15.1
elasticsearch==8.15.1
filelock==3.13.1
Flask==3.0.0
flask-cors==6.0.1
flask-talisman==1.1.0
Flask-WTF==1.2.1
frozenlist==1.4.1
google-api-core==2.21.0
google-auth==2.35.0
google-cloud-aiplatform==1.70.0
google-cloud-bigquery==3.26.0
google-cloud-core==2.4.1
google-cloud-resource-manager==1.12.5
google-cloud-storage==2.18.2
google-crc32c==1.6.0
google-genai==1.52.0
google-resumable-media==2.7.2
googleapis-common-protos==1.65.0
grpc-google-iam-v1==0.13.1
grpcio==1.67.0
grpcio-status==1.67.0
gunicorn==21.2.0
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.10
iniconfig==2.0.0
itsdangerous==2.1.2
Jinja2==3.1.3
jiter==0.6.1
json-logging==1.3.0
jsonpatch==1.33
jsonpointer==3.0.0
langchain==0.3.27
langchain-core==0.3.75
langchain-openai==0.3.8
langchain-text-splitters==0.3.11
langsmith==0.4.21
MarkupSafe==2.1.3
msal==1.31.0
msal-extensions==1.2.0
multidict==6.1.0
mypy-extensions==1.0.0
numpy==1.26.4
openai==1.65.5
orjson==3.10.7
ormsgpack==1.12.0
packaging==23.2
pathspec==0.12.1
pipenv==2023.11.15
platformdirs==4.1.0
pluggy==1.3.0
portalocker==2.10.1
propcache==0.2.0
proto-plus==1.24.0
protobuf==5.28.2
pyasn1==0.6.1
pyasn1_modules==0.4.1
pycparser==2.22
pydantic==2.9.2
pydantic_core==2.23.4
PyJWT==2.9.0
pytest==7.4.4
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-jose==3.3.0
pytz==2024.2
PyYAML==6.0.2
regex==2024.9.11
requests==2.32.3
requests-toolbelt==1.0.0
rsa==4.9
shapely==2.0.6
six==1.16.0
sniffio==1.3.1
SQLAlchemy==2.0.36
tenacity==8.5.0
tiktoken==0.8.0
tqdm==4.66.5
typing-inspection==0.4.2
typing_extensions==4.12.2
urllib3==2.2.3
virtualenv==20.25.0
virtualenv-clone==0.5.7
websockets==15.0.1
Werkzeug==3.1.3
WTForms==3.1.2
xxhash==3.6.0
yarl==1.15.4
zstandard==0.25.0
